chore(ci): moved tests to tests/ directory and reorganized by their target
