+++ tests/components/SetupRepositoryAzure.test.jsx
@@ -1,7 +1,7 @@
 import { render, act } from "@testing-library/react";
 import React from "react";
-import { SetupRepositoryAzure } from "../components/SetupRepositoryAzure";
-import { changeControlValue } from "./testutils";
+import { SetupRepositoryAzure } from "../../src/components/SetupRepositoryAzure";
+import { changeControlValue } from "../testutils";
 
 it("can set fields", async () => {
   let ref = React.createRef();
