+++ tests/components/SetupRepositoryFilesystem.test.jsx
@@ -1,18 +1,18 @@
 import { render, act } from "@testing-library/react";
 import React from "react";
-import { SetupRepositoryToken } from "../components/SetupRepositoryToken";
-import { changeControlValue } from "./testutils";
+import { SetupRepositoryFilesystem } from "../../src/components/SetupRepositoryFilesystem";
+import { changeControlValue } from "../testutils";
 
 it("can set fields", async () => {
   let ref = React.createRef();
-  const { getByTestId } = render(<SetupRepositoryToken ref={ref} />);
+  const { getByTestId } = render(<SetupRepositoryFilesystem ref={ref} />);
 
   act(() => expect(ref.current.validate()).toBe(false));
   // required
-  changeControlValue(getByTestId("control-token"), "some-token");
+  changeControlValue(getByTestId("control-path"), "some-path");
   expect(ref.current.validate()).toBe(true);
 
   expect(ref.current.state).toStrictEqual({
-    token: "some-token",
+    path: "some-path",
   });
 });
