+++ tests/components/SetupRepositoryToken.test.jsx
@@ -1,18 +1,18 @@
 import { render, act } from "@testing-library/react";
 import React from "react";
-import { SetupRepositoryFilesystem } from "../components/SetupRepositoryFilesystem";
-import { changeControlValue } from "./testutils";
+import { SetupRepositoryToken } from "../../src/components/SetupRepositoryToken";
+import { changeControlValue } from "../testutils";
 
 it("can set fields", async () => {
   let ref = React.createRef();
-  const { getByTestId } = render(<SetupRepositoryFilesystem ref={ref} />);
+  const { getByTestId } = render(<SetupRepositoryToken ref={ref} />);
 
   act(() => expect(ref.current.validate()).toBe(false));
   // required
-  changeControlValue(getByTestId("control-path"), "some-path");
+  changeControlValue(getByTestId("control-token"), "some-token");
   expect(ref.current.validate()).toBe(true);
 
   expect(ref.current.state).toStrictEqual({
-    path: "some-path",
+    token: "some-token",
   });
 });
