+++ tests/components/SetupRepositoryWebDAV.test.jsx
@@ -1,7 +1,7 @@
 import { render, act } from "@testing-library/react";
 import React from "react";
-import { SetupRepositoryWebDAV } from "../components/SetupRepositoryWebDAV";
-import { changeControlValue } from "./testutils";
+import { SetupRepositoryWebDAV } from "../../src/components/SetupRepositoryWebDAV";
+import { changeControlValue } from "../testutils";
 
 it("can set fields", async () => {
   let ref = React.createRef();
