+++ tests/components/notifications/EmailNotificationMethod.test.jsx
@@ -1,7 +1,7 @@
 import { render, act } from "@testing-library/react";
 import React from "react";
-import { EmailNotificationMethod } from "../../components/notifications/EmailNotificationMethod";
-import { changeControlValue } from "../testutils";
+import { EmailNotificationMethod } from "../../../src/components/notifications/EmailNotificationMethod";
+import { changeControlValue } from "../../testutils";
 
 it("can set fields", async () => {
   let ref = React.createRef();
